requirements.txt
asgiref==3.2.10
Django==3.0
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.10.0
Pillow==7.2.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
